+++ ugit.GitHubWorkflow.PSDevOps.ps1
@@ -1,7 +1,12 @@
 ﻿#requires -Module PSDevOps
 Push-Location $PSScriptRoot
 Import-BuildStep -ModuleName ugit
-New-GitHubWorkflow -Name "Analyze, Test, Tag, and Publish" -On Push, PullRequest, Demand -Job PowerShellStaticAnalysis, TestPowerShellOnLinux, TagReleaseAndPublish, buildugit |
-    Set-Content .\.github\workflows\TestAndPublish.yml -Encoding UTF8 -PassThru
+New-GitHubWorkflow -Name "Analyze, Test, Tag, and Publish" -On Push, PullRequest, Demand -Job PowerShellStaticAnalysis, 
+    TestPowerShellOnLinux, 
+    TagReleaseAndPublish, 
+    buildugit -OutputPath .\.github\workflows\TestAndPublish.yml
+
+New-GitHubWorkflow -On Issue, 
+    Demand -Job RunGitPub -Name OnIssueChanged -OutputPath .\.github\workflows\OnIssue.yml
 
 Pop-Location